fix: 노래 년도 정보 내려주기 위한 수정

+++ src/main/java/com/curioussong/alsongdalsong/game/domain/RoomInfo.java
@@ -18,7 +18,7 @@ public class RoomInfo {
     private int maxGameRound;
     private int skipCount;
     private boolean answered;
-    private List<Integer> selectedYears;
+    private List<Long> selectedYears;
 
     private Map<Integer, Pair<GameMode, Song>> roundToInfo;
